Explicity cast otp_offset

+++ OpenCL/m18100_a0-pure.cl
@@ -60,7 +60,7 @@ __kernel void m18100_mxx (__global pw_t *pws, __constant const kernel_rule_t *ru
     sha1_hmac_final (&ctx);
 
     // calculate the offset using the least 4 bits of the last byte of our hash
-    const int otp_offset = ctx.opad.h[4] & 0xf;
+    const int otp_offset = (int) ctx.opad.h[4] & 0xf;
 
     // initialize a buffer for the otp code
     unsigned int otp_code = 0;
@@ -153,7 +153,7 @@ __kernel void m18100_sxx (__global pw_t *pws, __constant const kernel_rule_t *ru
     sha1_hmac_final (&ctx);
 
     // calculate the offset using the least 4 bits of the last byte of our hash
-    const int otp_offset = ctx.opad.h[4] & 0xf;
+    const int otp_offset = (int) ctx.opad.h[4] & 0xf;
 
     // initialize a buffer for the otp code
     unsigned int otp_code = 0;
